refactor: reorganize imports and move db initialization to __init__.py
